+++ packages/core/src/components/ozon-content/ozon-content.template.ts
@@ -2,10 +2,11 @@ import { html } from 'lit-html';
 
 import { OzonContent } from '@dso-toolkit/sources';
 
-export function ozonContentTemplate({ content, onAnchorClick }: OzonContent) {
+export function ozonContentTemplate({ content, inline, onAnchorClick }: OzonContent) {
   return html`
     <dso-ozon-content
       .content=${content}
+      ?inline=${inline}
       @anchorClick=${onAnchorClick}
     ></dso-ozon-content>
   `;